@@ -1,114 +1,153 @@
 {
   "name": "mapbox-gl",
   "description": "A WebGL interactive maps library",
-  "version": "0.24.0",
-  "main": "js/mapbox-gl.js",
-  "license": "BSD-3-Clause",
+  "version": "0.53.0",
+  "main": "dist/mapbox-gl.js",
+  "style": "dist/mapbox-gl.css",
+  "license": "SEE LICENSE IN LICENSE.txt",
   "repository": {
     "type": "git",
     "url": "git://github.com/mapbox/mapbox-gl-js.git"
   },
   "engines": {
-    "node": ">=4.0.0"
+    "node": ">=6.4.0"
   },
   "dependencies": {
-    "csscolorparser": "^1.0.2",
-    "earcut": "^2.0.3",
-    "feature-filter": "^2.2.0",
-    "geojson-rewind": "^0.1.0",
-    "geojson-vt": "^2.4.0",
-    "gl-matrix": "^2.3.1",
-    "grid-index": "^1.0.0",
-    "mapbox-gl-function": "^1.2.1",
-    "mapbox-gl-shaders": "mapbox/mapbox-gl-shaders#df162476980d9ee2ab6f8d0cf5a06e27aac60472",
-    "mapbox-gl-style-spec": "mapbox/mapbox-gl-style-spec#a6c95d33ff5ced2c0d7df995fd89eb557c0a353c",
-    "mapbox-gl-supported": "^1.2.0",
-    "package-json-versionify": "^1.0.2",
-    "pbf": "^1.3.2",
-    "pngjs": "^2.2.0",
-    "point-geometry": "^0.0.0",
-    "quickselect": "^1.0.0",
-    "request": "^2.39.0",
-    "shelf-pack": "^1.0.0",
-    "supercluster": "^2.0.1",
-    "tinyqueue": "^1.1.0",
-    "unassertify": "^2.0.0",
-    "unitbezier": "^0.0.0",
-    "vector-tile": "^1.3.0",
-    "vt-pbf": "^2.0.2",
-    "webworkify": "^1.4.0",
-    "whoots-js": "^2.0.0"
+    "@mapbox/geojson-rewind": "^0.4.0",
+    "@mapbox/geojson-types": "^1.0.2",
+    "@mapbox/jsonlint-lines-primitives": "^2.0.2",
+    "@mapbox/mapbox-gl-supported": "^1.4.0",
+    "@mapbox/point-geometry": "^0.1.0",
+    "@mapbox/tiny-sdf": "^1.1.0",
+    "@mapbox/unitbezier": "^0.0.0",
+    "@mapbox/vector-tile": "^1.3.1",
+    "@mapbox/whoots-js": "^3.1.0",
+    "csscolorparser": "~1.0.2",
+    "earcut": "^2.1.5",
+    "esm": "^3.0.84",
+    "geojson-vt": "^3.2.1",
+    "gl-matrix": "^3.0.0",
+    "grid-index": "^1.1.0",
+    "minimist": "0.0.8",
+    "murmurhash-js": "^1.0.0",
+    "pbf": "^3.0.5",
+    "potpack": "^1.0.1",
+    "quickselect": "^2.0.0",
+    "rw": "^1.3.3",
+    "supercluster": "^6.0.1",
+    "tinyqueue": "^2.0.0",
+    "vt-pbf": "^3.1.1"
   },
   "devDependencies": {
-    "babel-preset-react": "^6.11.1",
-    "babelify": "^7.3.0",
+    "@mapbox/batfish": "^1.9.4",
+    "@mapbox/flow-remove-types": "^1.3.0-await.upstream.2",
+    "@mapbox/mapbox-gl-rtl-text": "^0.2.0",
+    "@mapbox/mapbox-gl-test-suite": "file:test/integration",
+    "@octokit/rest": "^15.15.1",
+    "babel-eslint": "^10.0.1",
     "benchmark": "~2.1.0",
-    "browserify": "^13.0.0",
-    "clipboard": "^1.5.12",
-    "concat-stream": "1.5.1",
-    "coveralls": "^2.11.8",
-    "doctrine": "^1.2.1",
-    "documentation": "https://github.com/documentationjs/documentation/archive/bb41619c734e59ef3fbc3648610032efcfdaaace.tar.gz",
-    "documentation-theme-utils": "3.0.0",
-    "envify": "^3.4.0",
-    "eslint": "^3.3.1",
-    "eslint-config-mourner": "^2.0.0",
-    "eslint-plugin-html": "^1.5.1",
-    "gl": "^4.0.1",
-    "handlebars": "4.0.5",
-    "highlight.js": "9.3.0",
-    "jsdom": "^9.4.2",
-    "json-loader": "^0.5.4",
-    "lodash": "^4.13.1",
-    "mapbox-gl-test-suite": "mapbox/mapbox-gl-test-suite#c736ea5acc1094488394f3fc5e5c1f6f94279c71",
-    "memory-fs": "^0.3.0",
-    "minifyify": "^7.0.1",
-    "npm-run-all": "^3.0.0",
-    "nyc": "^8.3.0",
-    "proxyquire": "^1.7.9",
-    "remark": "4.2.2",
-    "remark-html": "3.0.0",
-    "sinon": "^1.15.4",
+    "browserify": "^16.1.0",
+    "d3": "^4.12.0",
+    "documentation": "~8.1.1",
+    "ejs": "^2.5.7",
+    "eslint": "^5.8.0",
+    "eslint-config-mourner": "^3.0.0",
+    "eslint-plugin-flowtype": "^3.2.1",
+    "eslint-plugin-html": "^4.0.6",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-react": "^7.11.1",
+    "execcommand-copy": "^1.1.0",
+    "flow-bin": "^0.91.0",
+    "github-slugger": "^1.1.1",
+    "gl": "^4.1.1",
+    "glob": "^7.0.3",
+    "is-builtin-module": "^3.0.0",
+    "jsdom": "^13.0.0",
+    "json-stringify-pretty-compact": "^1.0.4",
+    "jsonwebtoken": "^8.3.0",
+    "mock-geolocation": "^1.0.11",
+    "npm-run-all": "^4.0.1",
+    "nyc": "^13.1.0",
+    "object.entries": "^1.0.4",
+    "pirates": "^3.0.2",
+    "pngjs": "^3.3.3",
+    "postcss-cli": "^5.0.0",
+    "postcss-inline-svg": "^3.1.1",
+    "pretty-bytes": "^5.1.0",
+    "prismjs": "^1.8.1",
+    "prop-types": "^15.6.0",
+    "raw-loader": "^0.5.1",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-helmet": "^5.2.0",
+    "remark": "^8.0.0",
+    "remark-html": "^5.0.1",
+    "remark-react": "^4.0.1",
+    "request": "^2.79.0",
+    "rollup": "^1.1.2",
+    "rollup-plugin-buble": "^0.19.6",
+    "rollup-plugin-commonjs": "^9.2.0",
+    "rollup-plugin-json": "^3.1.0",
+    "rollup-plugin-node-resolve": "4.0.0",
+    "rollup-plugin-replace": "^2.1.0",
+    "rollup-plugin-sourcemaps": "^0.4.2",
+    "rollup-plugin-terser": "^4.0.2",
+    "rollup-plugin-unassert": "^0.2.0",
+    "sinon": "^7.1.1",
+    "slugg": "^1.2.1",
     "st": "^1.2.0",
-    "tap": "^5.7.0",
-    "transform-loader": "^0.2.3",
-    "unist-util-visit": "1.1.0",
-    "vinyl": "1.1.1",
-    "vinyl-fs": "2.4.3",
-    "watchify": "^3.7.0"
-  },
-  "browserify": {
-    "transform": [
-      "package-json-versionify",
-      "unassertify"
-    ]
+    "stylelint": "^9.7.1",
+    "stylelint-config-standard": "^18.2.0",
+    "tap": "^12.0.1"
   },
   "browser": {
-    "./js/util/window.js": "./js/util/browser/window.js",
-    "./js/util/web_worker.js": "./js/util/browser/web_worker.js"
+    "./src/shaders/index.js": "./src/shaders/shaders.js",
+    "./src/util/window.js": "./src/util/browser/window.js",
+    "./src/util/web_worker.js": "./src/util/browser/web_worker.js"
   },
+  "esm": true,
   "scripts": {
-    "build-dev": "browserify js/mapbox-gl.js --debug --ignore-transform unassertify --standalone mapboxgl > dist/mapbox-gl-dev.js && tap --no-coverage test/build/dev.test.js",
-    "watch-dev": "watchify js/mapbox-gl.js --debug --ignore-transform unassertify --standalone mapboxgl --outfile dist/mapbox-gl-dev.js --verbose",
-    "build-min": "browserify js/mapbox-gl.js --debug --plugin [minifyify --map mapbox-gl.js.map --output dist/mapbox-gl.js.map] --standalone mapboxgl > dist/mapbox-gl.js && tap --no-coverage test/build/min.test.js",
-    "build-token": "browserify debug/access_token.js --debug --transform envify > debug/access_token_generated.js",
-    "build-benchmarks": "BENCHMARK_VERSION=\"$(git rev-parse --abbrev-ref HEAD) $(git rev-parse --short=7 HEAD)\" browserify bench/benchmarks.js --plugin [minifyify --no-map] --transform envify --outfile bench/benchmarks_generated.js --verbose",
-    "watch-benchmarks": "BENCHMARK_VERSION=\"$(git rev-parse --abbrev-ref HEAD) $(git rev-parse --short=7 HEAD)\" watchify bench/benchmarks.js --plugin [minifyify --no-map] --transform envify --outfile bench/benchmarks_generated.js --verbose",
-    "watch-benchmarks-view": "watchify bench/benchmarks_view.js --debug --transform [babelify --presets react] --outfile bench/benchmarks_view_generated.js --verbose",
-    "start-server": "st --no-cache --localhost --port 9966 --index index.html .",
-    "start": "run-p build-token watch-dev watch-benchmarks watch-benchmarks-view start-server",
-    "start-debug": "run-p build-token watch-dev start-server",
-    "start-bench": "run-p build-token watch-benchmarks watch-benchmarks-view start-server",
-    "build-docs": "documentation build --github --format html --config documentation.yml --theme ./docs/_theme --output docs/api/",
-    "build": "npm run build-docs # invoked by publisher when publishing docs on the mb-pages branch",
-    "start-docs": "npm run build-min && npm run build-docs && jekyll serve --watch",
-    "lint": "eslint  --ignore-path .gitignore js test bench docs/_posts/examples/*.html",
-    "open-changed-examples": "git diff --name-only mb-pages HEAD -- docs/_posts/examples/*.html | awk '{print \"http://127.0.0.1:4000/mapbox-gl-js/example/\" substr($0,33,length($0)-37)}' | xargs open",
-    "test": "run-s lint test-unit",
+    "build-dev": "rollup -c --environment BUILD:dev",
+    "watch-dev": "rollup -c --environment BUILD:dev --watch",
+    "build-prod": "rollup -c --environment BUILD:production",
+    "build-prod-min": "rollup -c --environment BUILD:production,MINIFY:true",
+    "build-flow-types": "cp build/mapbox-gl.js.flow dist/mapbox-gl.js.flow && cp build/mapbox-gl.js.flow dist/mapbox-gl-dev.js.flow",
+    "build-css": "postcss -o dist/mapbox-gl.css src/css/mapbox-gl.css",
+    "build-style-spec": "cd src/style-spec && npm run build && cd ../.. && mkdir -p dist/style-spec && cp src/style-spec/dist/* dist/style-spec",
+    "watch-css": "postcss --watch -o dist/mapbox-gl.css src/css/mapbox-gl.css",
+    "build-token": "node build/generate-access-token-script.js",
+    "build-benchmarks": "BENCHMARK_VERSION=${BENCHMARK_VERSION:-\"$(git rev-parse --abbrev-ref HEAD) $(git rev-parse --short=7 HEAD)\"} rollup -c bench/versions/rollup_config_benchmarks.js",
+    "watch-benchmarks": "BENCHMARK_VERSION=${BENCHMARK_VERSION:-\"$(git rev-parse --abbrev-ref HEAD) $(git rev-parse --short=7 HEAD)\"} rollup -c bench/versions/rollup_config_benchmarks.js --watch",
+    "watch-style-benchmarks": "BENCHMARK_VERSION=${BENCHMARK_VERSION:-\"$(git rev-parse --abbrev-ref HEAD) $(git rev-parse --short=7 HEAD)\"} rollup -c bench/styles/rollup_config_benchmarks.js --watch",
+    "start-server": "st --no-cache -H 0.0.0.0 --port 9966 --index index.html .",
+    "start": "run-p build-token watch-css watch-dev watch-benchmarks watch-style-benchmarks start-server",
+    "start-debug": "run-p build-token watch-css watch-dev start-server",
+    "start-bench": "run-p build-token watch-benchmarks watch-style-benchmarks start-server",
+    "build-docs": "documentation build --github --format json --config ./docs/documentation.yml --output docs/components/api.json src/index.js",
+    "build": "run-s build-docs && batfish build # invoked by publisher when publishing docs on the publisher-production branch",
+    "start-docs": "run-s build-prod-min build-css build-docs && DEPLOY_ENV=local batfish start",
+    "lint": "eslint --cache --ignore-path .gitignore src test bench docs docs/pages/example/*.html debug/*.html",
+    "lint-docs": "documentation lint src/index.js",
+    "lint-css": "stylelint 'src/css/mapbox-gl.css'",
+    "open-changed-examples": "git diff --name-only publisher-production HEAD -- docs/pages/example/*.html | awk '{print \"http://127.0.0.1:4000/mapbox-gl-js/example/\" substr($0,33,length($0)-37)}' | xargs open",
+    "test": "run-s lint lint-css test-flow test-unit",
     "test-suite": "run-s test-render test-query",
-    "test-unit": "tap --reporter dot --no-coverage test/js",
-    "test-render": "node test/render.test.js",
+    "test-suite-clean": "find test/integration/{render,query}-tests -mindepth 2 -type d -exec test -e \"{}/actual.png\" \\; -not \\( -exec test -e \"{}/style.json\" \\; \\) -print | xargs -t rm -r",
+    "test-unit": "build/run-tap --reporter classic --no-coverage test/unit",
+    "test-build": "build/run-tap --no-coverage test/build/**/*.test.js",
+    "test-render": "node --max-old-space-size=2048 test/render.test.js",
     "test-query": "node test/query.test.js",
-    "test-cov": "nyc --reporter=text-summary --cache run-s test-unit test-render test-query"
-  }
+    "test-expressions": "build/run-node test/expression.test.js",
+    "test-flow": "build/run-node build/generate-flow-typed-style-spec && flow .",
+    "test-cov": "nyc --require=@mapbox/flow-remove-types/register --reporter=text-summary --reporter=lcov --cache run-s test-unit test-expressions test-query test-render",
+    "prepublishOnly": "run-s build-flow-types build-dev build-prod-min build-prod build-css build-style-spec test-build",
+    "codegen": "build/run-node build/generate-style-code.js && build/run-node build/generate-struct-arrays.js"
+  },
+  "files": [
+    "build/",
+    "dist/",
+    "flow-typed/",
+    "src/",
+    ".flowconfig"
+  ]
 }
