@@ -1,28 +1,35 @@
-'use strict';
+// @flow
 
-var util = require('../../js/util/util');
-var mapboxgl = require('../../js/mapbox-gl');
+import Map from '../../src/ui/map';
 
-module.exports = function createMap(options) {
-    options = util.extend({width: 512, height: 512}, options);
+export default function (options: any): Promise<Map> {
+    return new Promise((resolve, reject) => {
+        const container = document.createElement('div');
+        container.style.width = `${options.width || 512}px`;
+        container.style.height = `${options.width || 512}px`;
+        container.style.margin = '0 auto';
+        container.style.display = 'none';
+        (document.body: any).appendChild(container);
 
-    var element = document.createElement('div');
-    element.style.width = options.width + 'px';
-    element.style.height = options.height + 'px';
-    element.style.margin = '0 auto';
-    document.body.appendChild(element);
+        const map = new Map(Object.assign({
+            container,
+            style: 'mapbox://styles/mapbox/streets-v10'
+        }, options));
 
-    mapboxgl.accessToken = require('./access_token');
+        map
+            .on('load', () => {
+                // Stub out `_rerender`; benchmarks need to be the only trigger of `_render` from here on out.
+                map._rerender = () => {};
 
-    var map = new mapboxgl.Map(util.extend({
-        container: element,
-        style: 'mapbox://styles/mapbox/streets-v9',
-        interactive: false
-    }, options));
+                // If there's a pending rerender, cancel it.
+                if (map._frame) {
+                    map._frame.cancel();
+                    map._frame = null;
+                }
 
-    map.on('remove', function() {
-        map.getContainer().remove();
+                resolve(map);
+            })
+            .on('error', (e) => reject(e.error))
+            .on('remove', () => container.remove());
     });
-
-    return map;
-};
+}
