@@ -1,246 +1,429 @@
-'use strict';
-/*eslint no-unused-vars: ["error", { "varsIgnorePattern": "BenchmarksView|clipboard" }]*/
+import React from 'react';
+import ReactDOM from 'react-dom';
+import * as d3 from 'd3';
+import Axis from './lib/axis';
+import { kde, probabilitiesOfSuperiority } from './lib/statistics';
 
-var Clipboard = require('clipboard');
+const versionColor = d3.scaleOrdinal(['#1b9e77', '#7570b3', '#d95f02']);
+const formatSample = d3.format(".3r");
 
-// Benchmark results seem to be more consistent with a warmup and cooldown
-// period. These values are measured in milliseconds.
-var benchmarkCooldownTime = 250;
-var benchmarkWarmupTime  = 250;
-
-var BenchmarksView = React.createClass({
-
-    render: function() {
-        return <div style={{width: 960, paddingBottom: window.innerHeight, margin: '0 auto'}}>
-            {this.renderSidebarBenchmarks()}
-            {this.renderBenchmarks()}
-        </div>;
-    },
-
-    renderSidebarBenchmarks: function() {
-        return <div style={{paddingTop: 40, width: 280, position: 'fixed'}} className='text-right'>
-            <h1 className="space-bottom">Benchmarks</h1>
-            <div className="space-bottom small">
-                {Object.keys(this.state.results).map(this.renderSidebarBenchmark)}
-            </div>
-            <a
-                    className={[
-                        'icon',
-                        'clipboard',
-                        'button',
-                        (this.getStatus() === 'ended' ? '' : 'disabled')
-                    ].join(' ')}
-                    data-clipboard-text={this.renderTextBenchmarks()}>
-                Copy Results
-            </a>
-        </div>;
-    },
-
-    renderSidebarBenchmark: function(name) {
-        return <div
-                key={name}
-                className={[
-                    'space-bottom',
-                    this.getBenchmarkStatus(name) === 'waiting' ? 'quiet' : ''
-                ].join(' ')}>
-            <h3>{name}</h3>
-            {Object.keys(this.state.results[name]).map(this.renderSidebarBenchmarkVersion.bind(this, name))}
-        </div>;
-    },
-
-    renderSidebarBenchmarkVersion: function(name, version) {
-        var results = this.state.results[name][version];
-        var that = this;
-
-        return <div
-                onClick={function() {
-                    that.scrollToBenchmark(name, version);
-                }}
-                style={{cursor: 'pointer'}}
-                key={version}
-                className={results.status === 'waiting' ? 'quiet' : ''}>
-            <strong>{version}:</strong> {results.message || '...'}
-        </div>;
-    },
-
-    renderTextBenchmarks: function() {
-        var output = '# Benchmarks\n';
-        for (var name in this.state.results) {
-            output += '\n## ' + name + '\n\n';
-            for (var version in this.state.results[name]) {
-                var result = this.state.results[name][version];
-                output += '**' + version + ':** ' + (result.message || '...') + '\n';
-            }
-        }
-        return output;
-    },
-
-    renderBenchmarks: function() {
-        return <div style={{width: 590, marginLeft: 320, marginBottom: 60}}>
-            {Object.keys(this.state.results).map(this.renderBenchmark)}
-        </div>;
-    },
-
-    renderBenchmark: function(name) {
-        return <div key={name}>
-            {Object.keys(this.state.results[name]).map(this.renderBenchmarkVersion.bind(this, name))}
-        </div>;
-    },
-
-    renderBenchmarkVersion: function(name, version) {
-        var results = this.state.results[name][version];
-        return (
-                <div
-                    style={{paddingTop: 40}}
-                    id={name + version}
-                    key={version}
-                    className={results.status === 'waiting' ? 'quiet' : ''}>
-
-                    <h2 className='space-bottom'>{name} on {version}</h2>
-                {results.logs.map(function(log, index) {
-                    return <div key={index} className={'pad1 dark fill-' + log.color}>{log.message}</div>;
-                })}
-            </div>
-        );
-    },
-
-    scrollToBenchmark: function(name, version) {
-        var duration = 300;
-        var startTime = (new Date()).getTime();
-        var startYOffset = window.pageYOffset;
-
-        requestAnimationFrame(function frame() {
-            var endYOffset = document.getElementById(name + version).offsetTop;
-            var time = (new Date()).getTime();
-            var yOffset = Math.min((time - startTime) / duration, 1) * (endYOffset - startYOffset) + startYOffset;
-            window.scrollTo(0, yOffset);
-            if (time < startTime + duration) requestAnimationFrame(frame);
-        });
-    },
-
-    getInitialState: function() {
-        var results = {};
-
-        for (var name in this.props.benchmarks) {
-            for (var version in this.props.benchmarks[name]) {
-                if (!this.props.benchmarkFilter || this.props.benchmarkFilter(name, version)) {
-                    results[name] = results[name] || {};
-                    results[name][version] = {
-                        status: 'waiting',
-                        logs: []
-                    };
-                }
-            }
-        }
-
-        return { results: results };
-    },
-
-    componentDidMount: function() {
-        var that = this;
-
-        asyncSeries(Object.keys(that.state.results), function(name, callback) {
-            asyncSeries(Object.keys(that.state.results[name]), function(version, callback) {
-                that.scrollToBenchmark(name, version);
-                that.runBenchmark(name, version, callback);
-            }, callback);
-        }, function(err) {
-            if (err) throw err;
-        });
-    },
-
-    runBenchmark: function(name, version, outerCallback) {
-        var that = this;
-        var results = this.state.results[name][version];
-
-        function log(color, message) {
-            results.logs.push({
-                color: color || 'blue',
-                message: message
-            });
-            that.forceUpdate();
-        }
-
-        function callback() {
-            setTimeout(outerCallback, benchmarkCooldownTime);
-        }
-
-        results.status = 'running';
-        this.scrollToBenchmark(name, version);
-        log('dark', 'starting');
-
-        setTimeout(function() {
-            var emitter = that.props.benchmarks[name][version]();
-
-            emitter.on('log', function(event) {
-                log(event.color, event.message);
-
-            });
-
-            emitter.on('end', function(event) {
-                results.message = event.message;
-                results.status = 'ended';
-                log('green', event.message);
-                callback();
-
-            });
-
-            emitter.on('error', function(event) {
-                results.status = 'errored';
-                log('red', event.error);
-                callback();
-            });
-
-        }, benchmarkWarmupTime);
-    },
-
-    getBenchmarkVersionStatus: function(name, version) {
-        return this.state.results[name][version].status;
-    },
-
-    getBenchmarkStatus: function(name) {
-        return reduceStatuses(Object.keys(this.state.results[name]).map(function(version) {
-            return this.getBenchmarkVersionStatus(name, version);
-        }, this));
-    },
-
-    getStatus() {
-        return reduceStatuses(Object.keys(this.state.results).map(function(name) {
-            return this.getBenchmarkStatus(name);
-        }, this));
+class StatisticsPlot extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {width: 100};
     }
-});
 
-function reduceStatuses(statuses) {
-    if (statuses.indexOf('running') !== -1) {
-        return 'running';
-    } else if (statuses.indexOf('waiting') !== -1) {
-        return 'waiting';
-    } else {
-        return 'ended';
+    render() {
+        const margin = {top: 0, right: 20, bottom: 20, left: 0};
+        const width = this.state.width - margin.left - margin.right;
+        const height = 400 - margin.top - margin.bottom;
+        const kdeWidth = 100;
+
+        const summaries = this.props.versions
+            .filter(v => v.status === 'ended')
+            .map(v => v.summary);
+
+        const t = d3.scaleLinear()
+            .domain([
+                d3.min(summaries.map(s => s.min)),
+                d3.max(summaries.map(s => Math.min(s.max, s.q2 + 3 * s.iqr)))
+            ])
+            .range([height, 0])
+            .clamp(true)
+            .nice();
+
+        const b = d3.scaleBand()
+            .domain(this.props.versions.map(v => v.name))
+            .range([kdeWidth + 20, width])
+            .paddingOuter(0.15)
+            .paddingInner(0.3);
+
+        const versions = this.props.versions.map(v => ({
+            name: v.name,
+            samples: v.samples,
+            summary: v.summary,
+            density: kde(v.samples, v.summary, t.ticks(50))
+        }));
+
+        const p = d3.scaleLinear()
+            .domain([0, d3.max(versions.map(v => d3.max(v.density, d => d[1])))])
+            .range([0, kdeWidth]);
+
+        const line = d3.line()
+            .curve(d3.curveBasis)
+            .y(d => t(d[0]))
+            .x(d => p(d[1]));
+
+        return (
+            <svg
+                width="100%"
+                height={height + margin.top + margin.bottom}
+                style={{overflow: 'visible'}}
+                ref={(ref) => { this.ref = ref; }}>
+                <defs>
+                    <g id="up-arrow">
+                        <path transform="translate(-6, -2)" style={{stroke: "inherit", fill: "inherit"}}
+                            d="M2,10 L6,2 L10,10"></path>
+                    </g>
+                </defs>
+                <g transform={`translate(${margin.left},${margin.top})`}>
+                    <Axis orientation="bottom" scale={p} ticks={[2, "%"]} transform={`translate(0,${height})`}>
+                    </Axis>
+                    <Axis orientation="left" scale={t} tickFormat={formatSample}>
+                        <text fill='#000' textAnchor="end"  y={6} transform="rotate(-90)" dy=".71em">Time (ms)</text>
+                    </Axis>
+                    {versions.map((v, i) => {
+                        if (v.samples.length === 0)
+                            return null;
+
+                        const bandwidth = b.bandwidth();
+                        const color = versionColor(v.name);
+                        const scale = d3.scaleLinear()
+                            .domain([0, v.samples.length])
+                            .range([0, bandwidth]);
+
+                        const {
+                            mean,
+                            trimmedMean,
+                            q1,
+                            q2,
+                            q3,
+                            min,
+                            max,
+                            argmin,
+                            argmax
+                        } = v.summary;
+
+                        const tMax = t.domain()[1];
+
+                        return <g key={i}>
+                            <path
+                                fill="none"
+                                stroke={color}
+                                strokeWidth={2}
+                                strokeOpacity={0.7}
+                                d={line(v.density)} />
+                            <g transform={`translate(${b(v.name)},0)`}>
+                                {v.samples.map((d, i) =>
+                                    <circle
+                                        key={i}
+                                        fill={color}
+                                        cx={scale(i)}
+                                        cy={t(d)}
+                                        r={i === argmin || i === argmax ? 2 : 1}
+                                        style={{
+                                            fillOpacity: d < tMax ? 1 : 0
+                                        }}
+                                    />
+                                )}
+                                {v.samples.filter(d => d >= tMax)
+                                    .map((d, i) =>
+                                        <use key={i}
+                                            href="#up-arrow"
+                                            x={scale(i)}
+                                            y={t(d)}
+                                            style={{
+                                                stroke:color,
+                                                strokeWidth: i === argmin || i === argmax ? 2 : 1,
+                                                fill: 'rgba(200, 0, 0, 0.5)'
+                                            }}
+                                        />
+                                    )
+                                }
+                                <line // quartiles
+                                    x1={bandwidth / 2}
+                                    x2={bandwidth / 2}
+                                    y1={t(q1)}
+                                    y2={t(q3)}
+                                    stroke={color}
+                                    strokeWidth={bandwidth}
+                                    strokeOpacity={0.5} />
+                                <line // median
+                                    x1={bandwidth / 2}
+                                    x2={bandwidth / 2}
+                                    y1={t(q2) - 0.5}
+                                    y2={t(q2) + 0.5}
+                                    stroke={color}
+                                    strokeWidth={bandwidth}
+                                    strokeOpacity={1} />
+                                <use href="#up-arrow" // mean
+                                    style={{ stroke: color, fill: color, fillOpacity: 0.4 }}
+                                    transform={mean >= tMax ? 'translate(-10, 0)' : `translate(-5, ${t(mean)}) rotate(90)`}
+                                    x={0}
+                                    y={0} />
+                                <use href="#up-arrow" // trimmed mean
+                                    style={{ stroke: color, fill: color }}
+                                    transform={`translate(-5, ${t(trimmedMean)}) rotate(90)`}
+                                    x={0}
+                                    y={0} />
+                                {[mean, trimmedMean].map((d, i) =>
+                                    <text // left
+                                        key={i}
+                                        dx={-16}
+                                        dy='.3em'
+                                        x={0}
+                                        y={t(d)}
+                                        textAnchor='end'
+                                        fontSize={10}
+                                        fontFamily='sans-serif'>{formatSample(d)}</text>
+                                )}
+                                {[[argmin, min], [argmax, max]].map((d, i) =>
+                                    <text // extent
+                                        key={i}
+                                        dx={0}
+                                        dy={i === 0 ? '1.3em' : '-0.7em'}
+                                        x={scale(d[0])}
+                                        y={t(d[1])}
+                                        textAnchor='middle'
+                                        fontSize={10}
+                                        fontFamily='sans-serif'>{formatSample(d[1])}</text>
+                                )}
+                                {[q1, q2, q3].map((d, i) =>
+                                    <text // right
+                                        key={i}
+                                        dx={6}
+                                        dy='.3em'
+                                        x={bandwidth}
+                                        y={t(d)}
+                                        textAnchor='start'
+                                        fontSize={10}
+                                        fontFamily='sans-serif'>{formatSample(d)}</text>
+                                )}
+                            </g>
+                        </g>;
+                    })}
+                </g>
+            </svg>
+        );
+    }
+
+    componentDidMount() {
+        this.setState({ width: this.ref.clientWidth });
     }
 }
 
-var clipboard = new Clipboard('.clipboard');
-
-ReactDOM.render(
-    <BenchmarksView
-        benchmarks={window.mapboxglBenchmarks}
-        benchmarkFilter={function(name) {
-            var nameFilter = window.location.hash.substr(1);
-            return !nameFilter || name === nameFilter;
-        }}
-    />,
-    document.getElementById('benchmarks')
-);
-
-function asyncSeries(array, iterator, callback) {
-    if (array.length) {
-        iterator(array[0], function(err) {
-            if (err) callback(err);
-            else asyncSeries(array.slice(1), iterator, callback);
-        });
-    } else {
-        callback();
+class RegressionPlot extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {width: 100};
     }
+
+    render() {
+        const margin = {top: 10, right: 20, bottom: 30, left: 0};
+        const width = this.state.width - margin.left - margin.right;
+        const height = 200 - margin.top - margin.bottom;
+        const versions = this.props.versions.filter(version => version.regression);
+
+        const x = d3.scaleLinear()
+            .domain([0, d3.max(versions.map(version => d3.max(version.regression.data, d => d[0])))])
+            .range([0, width])
+            .nice();
+
+        const y = d3.scaleLinear()
+            .domain([0, d3.max(versions.map(version => d3.max(version.regression.data, d => d[1])))])
+            .range([height, 0])
+            .nice();
+
+        const line = d3.line()
+            .x(d => x(d[0]))
+            .y(d => y(d[1]));
+
+        return (
+            <svg
+                width="100%"
+                height={height + margin.top + margin.bottom}
+                style={{overflow: 'visible'}}
+                ref={(ref) => { this.ref = ref; }}>
+                <g transform={`translate(${margin.left},${margin.top})`}>
+                    <Axis orientation="bottom" scale={x} transform={`translate(0,${height})`}>
+                        <text fill='#000' textAnchor="end" y={-6} x={width}>Iterations</text>
+                    </Axis>
+                    <Axis orientation="left" scale={y} ticks={4} tickFormat={formatSample}>
+                        <text fill='#000' textAnchor="end"  y={6} transform="rotate(-90)" dy=".71em">Time (ms)</text>
+                    </Axis>
+                    {versions.map((v, i) =>
+                        <g
+                            key={i}
+                            fill={versionColor(v.name)}
+                            fillOpacity="0.7">
+                            {v.regression.data.map(([a, b], i) =>
+                                <circle key={i} r="2" cx={x(a)} cy={y(b)}/>
+                            )}
+                            <path
+                                stroke={versionColor(v.name)}
+                                strokeWidth={1}
+                                strokeOpacity={0.5}
+                                d={line(v.regression.data.map(d => [
+                                    d[0],
+                                    d[0] * v.regression.slope + v.regression.intercept
+                                ]))} />
+                        </g>
+                    )}
+                </g>
+            </svg>
+        );
+    }
+
+    componentDidMount() {
+        this.setState({ width: this.ref.clientWidth });
+    }
+}
+
+class BenchmarkStatistic extends React.Component {
+    render() {
+        switch (this.props.status) {
+        case 'waiting':
+            return <p className="quiet"></p>;
+        case 'running':
+            return <p>Running...</p>;
+        case 'error':
+            return <p>{this.props.error.message}</p>;
+        default:
+            return this.props.statistic(this.props);
+        }
+    }
+}
+
+class BenchmarkRow extends React.Component {
+    render() {
+        const endedCount = this.props.versions.filter(version => version.status === 'ended').length;
+
+        let master;
+        let current;
+        if (/master/.test(this.props.versions[0].name)) {
+            [master, current] = this.props.versions;
+        } else {
+            [current, master] = this.props.versions;
+        }
+
+        let change;
+        let pInferiority;
+        if (endedCount === 2) {
+            const delta = current.summary.trimmedMean - master.summary.trimmedMean;
+            // Use "Cohen's d" (modified to used the trimmed mean/sd) to decide
+            // how much to emphasize difference between means
+            // https://en.wikipedia.org/wiki/Effect_size#Cohen.27s_d
+            const pooledDeviation = Math.sqrt(
+                (
+                    (master.samples.length - 1) * Math.pow(master.summary.windsorizedDeviation, 2) +
+                    (current.samples.length - 1) * Math.pow(current.summary.windsorizedDeviation, 2)
+                ) /
+                (master.samples.length + current.samples.length - 2)
+            );
+            const d = delta / pooledDeviation;
+
+            const {superior, inferior} = probabilitiesOfSuperiority(master.samples, current.samples);
+
+            change = <span className={d < 0.2 ? 'quiet' : d < 1.5 ? '' : 'strong'}>(
+                {delta > 0 ? '+' : ''}{formatSample(delta)} ms / {d.toFixed(1)} std devs
+            )</span>;
+
+            const comparison = inferior > superior ? 'SLOWER' : 'faster';
+            const probability = Math.max(inferior, superior);
+            pInferiority = <p className={`center ${probability > 0.90 ? 'strong' : 'quiet'}`}>
+                {(probability * 100).toFixed(0)}%
+                chance that a random <svg width={8} height={8}><circle fill={versionColor(current.name)} cx={4} cy={4} r={4} /></svg> sample is
+                {comparison} than a random <svg width={8} height={8}><circle fill={versionColor(master.name)} cx={4} cy={4} r={4} /></svg> sample.
+            </p>;
+        }
+
+        return (
+            <div className="col12 clearfix space-bottom">
+                <table className="fixed space-bottom">
+                    <tbody>
+                        <tr><th><h2 className="col4"><a href={`#${this.props.name}`} onClick={this.reload}>{this.props.name}</a></h2></th>
+                            {this.props.versions.map(version => <th style={{color: versionColor(version.name)}} key={version.name}>{version.name}</th>)}</tr>
+                        {this.props.location && <tr>
+                            <th><p style={{color: '#1287A8'}}>{this.props.location.description}</p></th>
+                            <th><p style={{color: '#1287A8'}}>Zoom Level: {this.props.location.zoom}</p></th>
+                            <th><p style={{color: '#1287A8'}}>Lat: {this.props.location.center[1]} Lng: {this.props.location.center[0]}</p></th>
+                        </tr>}
+                        {this.renderStatistic('(20% trimmed) Mean',
+                            (version) => <p>
+                                {formatSample(version.summary.trimmedMean)} ms
+                                {current && version.name === current.name && change}
+                            </p>)}
+                        {this.renderStatistic('(Windsorized) Deviation',
+                            (version) => <p>{formatSample(version.summary.windsorizedDeviation)} ms</p>)}
+                        {this.renderStatistic('R² Slope / Correlation',
+                            (version) => <p>{formatSample(version.regression.slope)} ms / {version.regression.correlation.toFixed(3)} {
+                                version.regression.correlation < 0.9 ? '\u2620\uFE0F' :
+                                version.regression.correlation < 0.99 ? '\u26A0\uFE0F' : ''}</p>)}
+                        {this.renderStatistic('Minimum',
+                            (version) => <p>{formatSample(version.summary.min)} ms</p>)}
+                        {pInferiority && <tr><td colSpan={3}>{pInferiority}</td></tr>}
+                    </tbody>
+                </table>
+                {endedCount > 0 && <StatisticsPlot versions={this.props.versions}/>}
+                {endedCount > 0 && <RegressionPlot versions={this.props.versions}/>}
+            </div>
+        );
+    }
+
+    renderStatistic(title, statistic) {
+        return (
+            <tr>
+                <th>{title}</th>
+                {this.props.versions.map(version =>
+                    <td key={version.name}><BenchmarkStatistic statistic={statistic} {...version}/></td>
+                )}
+            </tr>
+        );
+    }
+
+    reload() {
+        location.reload();
+    }
+}
+
+class BenchmarksTable extends React.Component {
+    constructor(props) {
+        super(props);
+        this.state = {sharing: false};
+        this.share = this.share.bind(this);
+    }
+
+    render() {
+        return (
+            <div style={{width: 960, margin: '2em auto'}}>
+                {this.state.sharing && <span className='loading'></span>}
+                <h1 className="space-bottom1">Mapbox GL JS Benchmarks – {
+                    this.props.finished ?
+                        <span>Finished <button className='button fr icon share' onClick={this.share}>Share</button></span> :
+                        <span>Running</span>}</h1>
+                {this.props.benchmarks.map((benchmark, i) => {
+                    return <BenchmarkRow key={`${benchmark.name}-${i}`} {...benchmark}/>;
+                })}
+            </div>
+        );
+    }
+
+    share() {
+        document.querySelectorAll('script').forEach(e => e.remove());
+        const share = document.querySelector('.share');
+        share.style.display = 'none';
+
+        const body = JSON.stringify({
+            "public": true,
+            "files": {
+                "index.html": {
+                    "content": document.body.parentElement.outerHTML
+                }
+            }
+        });
+        this.setState({sharing: true});
+
+        fetch('https://api.github.com/gists', { method: 'POST', body })
+            .then(response => response.json())
+            .then(json => { window.location = `https://bl.ocks.org/anonymous/raw/${json.id}/`; });
+    }
+}
+
+export default function updateUI(benchmarks, finished) {
+    finished = !!finished;
+
+    ReactDOM.render(
+        <BenchmarksTable benchmarks={benchmarks} finished={finished}/>,
+        document.getElementById('benchmarks')
+    );
 }
