@@ -1,12 +1,17 @@
-'use strict';
+/* eslint-disable import/unambiguous, import/no-commonjs, no-global-assign */
 
-var querySuite = require('mapbox-gl-test-suite').query;
-var suiteImplementation = require('./suite_implementation');
+require('./stub_loader');
+require('@mapbox/flow-remove-types/register');
+require = require("esm")(module, true);
 
-var tests;
+const querySuite = require('./integration/lib/query');
+const suiteImplementation = require('./suite_implementation');
+const ignores = require('./ignores.json');
+
+let tests;
 
 if (process.argv[1] === __filename && process.argv.length > 2) {
     tests = process.argv.slice(2);
 }
 
-querySuite.run('js', {tests: tests}, suiteImplementation);
+querySuite.run('js', {tests, ignores}, suiteImplementation);
