@@ -1,75 +1,117 @@
-'use strict';
+import {PNG} from 'pngjs';
+import Map from '../src/ui/map';
+import config from '../src/util/config';
+import window from '../src/util/window';
+import browser from '../src/util/browser';
+import {plugin as rtlTextPlugin} from '../src/source/rtl_text_plugin';
+import rtlText from '@mapbox/mapbox-gl-rtl-text';
+import fs from 'fs';
+import path from 'path';
+import customLayerImplementations from './integration/custom_layer_implementations';
 
-var ajax =  require('../js/util/ajax');
-var sinon = require('sinon');
-var request = require('request');
-var PNG = require('pngjs').PNG;
-var Map = require('../js/ui/map');
-var window = require('../js/util/window');
+rtlTextPlugin['applyArabicShaping'] = rtlText.applyArabicShaping;
+rtlTextPlugin['processBidirectionalText'] = rtlText.processBidirectionalText;
+rtlTextPlugin['processStyledBidirectionalText'] = rtlText.processStyledBidirectionalText;
 
-module.exports = function(style, options, _callback) {
-    var wasCallbackCalled = false;
-    function callback() {
+module.exports = function(style, options, _callback) { // eslint-disable-line import/no-commonjs
+    let wasCallbackCalled = false;
+
+    const timeout = setTimeout(() => {
+        callback(new Error('Test timed out'));
+    }, options.timeout || 20000);
+
+    function callback(...args) {
         if (!wasCallbackCalled) {
+            clearTimeout(timeout);
             wasCallbackCalled = true;
-            _callback.apply(this, arguments);
+            _callback(...args);
         }
     }
 
     window.devicePixelRatio = options.pixelRatio;
 
-    var container = window.document.createElement('div');
-    container.offsetHeight = options.height;
-    container.offsetWidth = options.width;
+    if (options.addFakeCanvas) {
+        const fakeCanvas = createFakeCanvas(window.document, options.addFakeCanvas.id, options.addFakeCanvas.image);
+        window.document.body.appendChild(fakeCanvas);
+    }
 
-    var map = new Map({
-        container: container,
-        style: style,
+    const container = window.document.createElement('div');
+    Object.defineProperty(container, 'clientWidth', {value: options.width});
+    Object.defineProperty(container, 'clientHeight', {value: options.height});
+
+    // We are self-hosting test files.
+    config.REQUIRE_ACCESS_TOKEN = false;
+
+    const map = new Map({
+        container,
+        style,
         classes: options.classes,
         interactive: false,
         attributionControl: false,
-        preserveDrawingBuffer: true
+        preserveDrawingBuffer: true,
+        axonometric: options.axonometric || false,
+        skew: options.skew || [0, 0],
+        fadeDuration: options.fadeDuration || 0,
+        crossSourceCollisions: typeof options.crossSourceCollisions === "undefined" ? true : options.crossSourceCollisions
     });
 
     // Configure the map to never stop the render loop
     map.repaint = true;
 
+    let now = 0;
+    browser.now = function() {
+        return now;
+    };
+
     if (options.debug) map.showTileBoundaries = true;
-    if (options.collisionDebug) map.showCollisionBoxes = true;
     if (options.showOverdrawInspector) map.showOverdrawInspector = true;
 
-    var gl = map.painter.gl;
+    const gl = map.painter.context.gl;
 
-    map.once('load', function() {
-        applyOperations(map, options.operations, function() {
-            var w = options.width * window.devicePixelRatio;
-            var h = options.height * window.devicePixelRatio;
+    map.once('load', () => {
+        if (options.collisionDebug) {
+            map.showCollisionBoxes = true;
+            if (options.operations) {
+                options.operations.push(["wait"]);
+            } else {
+                options.operations = [["wait"]];
+            }
+        }
+        applyOperations(map, options.operations, () => {
+            const viewport = gl.getParameter(gl.VIEWPORT);
+            const w = viewport[2];
+            const h = viewport[3];
 
-            var pixels = new Uint8Array(w * h * 4);
+            const pixels = new Uint8Array(w * h * 4);
             gl.readPixels(0, 0, w, h, gl.RGBA, gl.UNSIGNED_BYTE, pixels);
 
-            var data = new Buffer(pixels);
+            const data = new Buffer(pixels);
 
             // Flip the scanlines.
-            var stride = w * 4;
-            var tmp = new Buffer(stride);
-            for (var i = 0, j = h - 1; i < j; i++, j--) {
-                var start = i * stride;
-                var end = j * stride;
+            const stride = w * 4;
+            const tmp = new Buffer(stride);
+            for (let i = 0, j = h - 1; i < j; i++, j--) {
+                const start = i * stride;
+                const end = j * stride;
                 data.copy(tmp, 0, start, start + stride);
                 data.copy(data, start, end, end + stride);
                 tmp.copy(data, end);
             }
 
-            var results = options.queryGeometry ?
-                map.queryRenderedFeatures(options.queryGeometry, options) :
+            const results = options.queryGeometry ?
+                map.queryRenderedFeatures(options.queryGeometry, options.queryOptions || {}) :
                 [];
 
             map.remove();
             gl.getExtension('STACKGL_destroy_context').destroy();
-            delete map.painter.gl;
+            delete map.painter.context.gl;
 
-            callback(null, data, results.map(function (feature) {
+            if (options.addFakeCanvas) {
+                const fakeCanvas = window.document.getElementById(options.addFakeCanvas.id);
+                fakeCanvas.parentNode.removeChild(fakeCanvas);
+            }
+
+            callback(null, data, results.map((feature) => {
                 feature = feature.toJSON();
                 delete feature.layer;
                 return feature;
@@ -77,115 +119,57 @@ module.exports = function(style, options, _callback) {
 
         });
     });
+
+    function applyOperations(map, operations, callback) {
+        const operation = operations && operations[0];
+        if (!operations || operations.length === 0) {
+            callback();
+
+        } else if (operation[0] === 'wait') {
+            if (operation.length > 1) {
+                now += operation[1];
+                map._render();
+                applyOperations(map, operations.slice(1), callback);
+
+            } else {
+                const wait = function() {
+                    if (map.loaded()) {
+                        applyOperations(map, operations.slice(1), callback);
+                    } else {
+                        map.once('render', wait);
+                    }
+                };
+                wait();
+            }
+
+        } else if (operation[0] === 'sleep') {
+            // Prefer "wait", which renders until the map is loaded
+            // Use "sleep" when you need to test something that sidesteps the "loaded" logic
+            setTimeout(() => {
+                applyOperations(map, operations.slice(1), callback);
+            }, operation[1]);
+        } else if (operation[0] === 'addImage') {
+            const {data, width, height} = PNG.sync.read(fs.readFileSync(path.join(__dirname, './integration', operation[2])));
+            map.addImage(operation[1], {width, height, data: new Uint8Array(data)}, operation[3] || {});
+            applyOperations(map, operations.slice(1), callback);
+        } else if (operation[0] === 'addCustomLayer') {
+            map.addLayer(new customLayerImplementations[operation[1]](), operation[2]);
+            map._render();
+            applyOperations(map, operations.slice(1), callback);
+
+        } else {
+            map[operation[0]](...operation.slice(1));
+            applyOperations(map, operations.slice(1), callback);
+        }
+    }
 };
 
-function applyOperations(map, operations, callback) {
-    var operation = operations && operations[0];
-    if (!operations || operations.length === 0) {
-        callback();
-
-    } else if (operation[0] === 'wait') {
-        var wait = function() {
-            if (map.loaded()) {
-                applyOperations(map, operations.slice(1), callback);
-            } else {
-                map.once('render', wait);
-            }
-        };
-        wait();
-
-    } else {
-        map[operation[0]].apply(map, operation.slice(1));
-        applyOperations(map, operations.slice(1), callback);
-    }
+function createFakeCanvas(document, id, imagePath) {
+    const fakeCanvas = document.createElement('canvas');
+    const image = PNG.sync.read(fs.readFileSync(path.join(__dirname, './integration', imagePath)));
+    fakeCanvas.id = id;
+    fakeCanvas.data = image.data;
+    fakeCanvas.width = image.width;
+    fakeCanvas.height = image.height;
+    return fakeCanvas;
 }
-
-function fakeImage(png) {
-    return {
-        width: png.width,
-        height: png.height,
-        data: png.data.slice(),
-        complete: true,
-        getData: function() { return this.data; }
-    };
-}
-
-var cache = {};
-
-function cached(data, callback) {
-    setImmediate(function () {
-        callback(null, data);
-    });
-}
-
-sinon.stub(ajax, 'getJSON', function(url, callback) {
-    if (cache[url]) return cached(cache[url], callback);
-    return request(url, function(error, response, body) {
-        if (!error && response.statusCode >= 200 && response.statusCode < 300) {
-            var data;
-            try {
-                data = JSON.parse(body);
-            } catch (err) {
-                return callback(err);
-            }
-            cache[url] = data;
-            callback(null, data);
-        } else {
-            callback(error || new Error(response.statusCode));
-        }
-    });
-});
-
-sinon.stub(ajax, 'getArrayBuffer', function(url, callback) {
-    if (cache[url]) return cached(cache[url], callback);
-    return request({url: url, encoding: null}, function(error, response, body) {
-        if (!error && response.statusCode >= 200 && response.statusCode < 300) {
-            var ab = new ArrayBuffer(body.length);
-            var view = new Uint8Array(ab);
-            for (var i = 0; i < body.length; ++i) {
-                view[i] = body[i];
-            }
-            cache[url] = ab;
-            callback(null, ab);
-        } else {
-            callback(error || new Error(response.statusCode));
-        }
-    });
-});
-
-sinon.stub(ajax, 'getImage', function(url, callback) {
-    if (cache[url]) return cached(fakeImage(cache[url]), callback);
-    return request({url: url, encoding: null}, function(error, response, body) {
-        if (!error && response.statusCode >= 200 && response.statusCode < 300) {
-            new PNG().parse(body, function(err, png) {
-                if (err) return callback(err);
-                cache[url] = png;
-                callback(null, fakeImage(png));
-            });
-        } else {
-            callback(error || new Error(response.statusCode));
-        }
-    });
-});
-
-// Hack: since node doesn't have any good video codec modules, just grab a png with
-// the first frame and fake the video API.
-sinon.stub(ajax, 'getVideo', function(urls, callback) {
-    return request({url: urls[0], encoding: null}, function(error, response, body) {
-        if (!error && response.statusCode >= 200 && response.statusCode < 300) {
-            new PNG().parse(body, function(err, png) {
-                if (err) return callback(err);
-                callback(null, {
-                    readyState: 4, // HAVE_ENOUGH_DATA
-                    addEventListener: function() {},
-                    play: function() {},
-                    width: png.width,
-                    height: png.height,
-                    data: png.data
-                });
-            });
-        } else {
-            callback(error || new Error(response.statusCode));
-        }
-    });
-});
